@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.41'
+    ext.kotlin_version = '1.5.10'
 
     repositories {
         mavenCentral()
@@ -11,55 +11,47 @@ buildscript {
 
 buildscript {
     repositories {
-        jcenter()
+        gradlePluginPortal()
     }
     dependencies {
-        classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.4'
+        classpath 'gradle.plugin.com.github.jengelman.gradle.plugins:shadow:7.0.0'
     }
 }
 
-
 group 'net.skripthub'
-version '1.3'
+version '1.4'
 
 apply plugin: 'com.github.johnrengelman.shadow'
 apply plugin: 'java'
 apply plugin: 'kotlin'
 
-sourceCompatibility = 1.8
-
 repositories {
     mavenCentral()
-    maven { url "https://hub.spigotmc.org/nexus/content/repositories/snapshots" }
-    maven { url "https://oss.sonatype.org/content/repositories/snapshots" }
-    maven { url "https://raw.githubusercontent.com/bensku/mvn-repo/master" }
+    maven {
+        url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
+        content {
+            includeGroup 'org.bukkit'
+            includeGroup 'org.spigotmc'
+        }
+    }
+    maven { url = 'https://oss.sonatype.org/content/repositories/snapshots' }
+    maven { url = 'https://oss.sonatype.org/content/repositories/central' }
     maven { url 'https://repo.destroystokyo.com/repository/maven-public/' }
-    maven { url 'http://maven.sk89q.com/repo' }
-    maven { url 'http://nexus.hc.to/content/repositories/pub_releases' }
+    maven { url 'https://papermc.io/repo/repository/maven-public/' }
     maven { url 'https://jitpack.io' }
 }
 
 dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation 'com.github.SkriptLang:Skript:2.6'
 
-    implementation 'com.github.SkriptLang:Skript:dev37c'
+    implementation group: 'org.spigotmc', name: 'spigot-api', version:'1.17-R0.1-SNAPSHOT'
 
-    compileOnly group: 'org.spigotmc', name: 'spigot-api', version:'1.12.2-R0.1-SNAPSHOT'
-    compileOnly group: 'org.bukkit', name: 'bukkit', version:'1.12.2-R0.1-SNAPSHOT'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+    testImplementation group: 'junit', name: 'junit', version: '4.12'
 }
 
 shadowJar {
     dependencies {
         include dependency("org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version")
     }
-    configurations = [project.configurations.compile]
     relocate 'kotlin', 'net.skripthub.docstool'
 }
