@@ -6,7 +6,8 @@ data class AddonData(var name: String,
                      var effects: MutableList<SyntaxData> = mutableListOf(),
                      var expressions: MutableList<SyntaxData> = mutableListOf(),
                      var types: MutableList<SyntaxData> = mutableListOf(),
-                     var functions: MutableList<SyntaxData> = mutableListOf()) {
+                     var functions: MutableList<SyntaxData> = mutableListOf(),
+                     var sections: MutableList<SyntaxData> = mutableListOf()) {
     fun sortLists() {
         events.sortBy { info -> info.name }
         conditions.sortBy { info -> info.name }
@@ -14,5 +15,6 @@ data class AddonData(var name: String,
         expressions.sortBy { info -> info.name }
         types.sortBy { info -> info.name }
         functions.sortBy { info -> info.name }
+        sections.sortBy { info -> info.name }
     }
 }
